refactor(Web-UI): Moved tests to new folder structure
