+++ src/AppHost/ClientApp/tests/nuxt/stores/account-store/create-plex-account.test.ts
@@ -1,4 +1,4 @@
-import { describe, beforeAll, test, expect } from 'vitest';
+import { describe, beforeAll, test, expect, beforeEach } from 'vitest';
 import { createPinia, setActivePinia } from 'pinia';
 import { PlexAccountPaths, PlexLibraryPaths, PlexServerPaths } from '@api-urls';
 import { baseSetup, baseVars, getAxiosMock, subscribeSpyTo } from '@services-test-base';
