+++ src/AppHost/ClientApp/tests/nuxt/stores/account-store/setup.test.ts
@@ -1,12 +1,12 @@
-import { describe, beforeAll, test, expect } from 'vitest';
+import { describe, beforeAll, test, expect, beforeEach } from 'vitest';
 import { createPinia, setActivePinia } from 'pinia';
 import { subscribeSpyTo, baseSetup, getAxiosMock, baseVars } from '@services-test-base';
-import { generateResultDTO } from '@mock';
-import { FolderPathPaths } from '@api/api-paths';
+import { PlexAccountPaths } from '@api/api-paths';
 import type { ISetupResult } from '@interfaces';
-import { useFolderPathStore } from '@store';
+import { generateResultDTO } from '@mock';
+import { useAccountStore } from '@store';
 
-describe('FolderPathStore.setup()', () => {
+describe('AccountStore.setup()', () => {
 	let { mock } = baseVars();
 
 	beforeAll(() => {
@@ -20,15 +20,16 @@ describe('FolderPathStore.setup()', () => {
 
 	test('Should return success and complete when setup is run', async () => {
 		// Arrange
-		const folderPathStore = useFolderPathStore();
-		mock.onGet(FolderPathPaths.getAllFolderPathsEndpoint()).reply(200, generateResultDTO([]));
+		const accountStore = useAccountStore();
+		mock.onGet(PlexAccountPaths.getAllPlexAccountsEndpoint()).reply(200, generateResultDTO([]));
+		const setup$ = accountStore.setup();
 		const setupResult: ISetupResult = {
 			isSuccess: true,
-			name: 'useFolderPathStore',
+			name: 'useAccountStore',
 		};
 
 		// Act
-		const result = subscribeSpyTo(folderPathStore.setup());
+		const result = subscribeSpyTo(setup$);
 		await result.onComplete();
 
 		// Assert
