+++ src/AppHost/ClientApp/tests/nuxt/stores/alert-store/setup.test.ts
