+++ src/AppHost/ClientApp/tests/nuxt/stores/background-jobs-store/getJobStatusUpdate.test.ts
@@ -1,4 +1,4 @@
-import { beforeAll, describe, expect, test } from 'vitest';
+import { beforeAll, describe, expect, test, beforeEach } from 'vitest';
 import { createPinia, setActivePinia } from 'pinia';
 import { baseSetup, baseVars, getAxiosMock, subscribeSpyTo } from '@services-test-base';
 import type { InspectPlexServerJobUpdateDTO, SyncServerMediaJobUpdateDTO } from '@dto';
