+++ src/AppHost/ClientApp/tests/nuxt/stores/download-store/setup.test.ts
@@ -1,4 +1,4 @@
-import { describe, beforeAll, test, expect } from 'vitest';
+import { describe, beforeAll, test, expect, beforeEach } from 'vitest';
 import { createPinia, setActivePinia } from 'pinia';
 import { subscribeSpyTo, baseSetup, getAxiosMock, baseVars } from '@services-test-base';
 import { generateResultDTO } from '@mock';
