+++ src/AppHost/ClientApp/tests/nuxt/stores/global-store/get-page-ready.test.ts
