+++ src/AppHost/ClientApp/tests/nuxt/stores/global-store/setup.test.ts
