+++ src/AppHost/ClientApp/tests/nuxt/stores/help-store/setup.test.ts
@@ -1,4 +1,4 @@
-import { describe, beforeAll, test, expect } from 'vitest';
+import { describe, beforeAll, test, expect, beforeEach } from 'vitest';
 import { createPinia, setActivePinia } from 'pinia';
 import { subscribeSpyTo, baseSetup, baseVars } from '@services-test-base';
 import type { ISetupResult } from '@interfaces';
