+++ src/AppHost/ClientApp/tests/nuxt/stores/media-store/setup.test.ts
