+++ src/AppHost/ClientApp/tests/nuxt/stores/server-store/setup.test.ts
