+++ src/AppHost/ClientApp/tests/nuxt/stores/settings-store/setup.test.ts
