+++ src/AppHost/ClientApp/tests/nuxt/stores/signalr-store/setup.test.ts
